Made responsive layout for mobile. Also removed lots of icons.
The layout was not rendering well for mobile devices. Still needs to be
tested but this was tested so far in a browser width that was minimum
(chrome).

There was a bunch of slowness on generating the static site and that was
because of all the extra icons. Add them as needed.